requirements.txt
openai migrate
praw
python-dotenv
tqdm
